requirements.txt
Django==1.11.6
django-bootstrap4==0.0.4
pythonwhois==2.4.3
requests==2.18.4
paramiko==2.3.1
django-su==0.5.2
rrdtool==0.1.11
hurry.filesize==0.9
python-dateutil==2.6.1
